Loading questions from interent file
I added the ability to be able to load questions from a JSON file from the internet. The files that I used were https://testing.sarajaksa.eu/list_of_questions.js and https://testing.sarajaksa.eu/flash_questions.js.

I added this, because in the future, we are expected to be able to push the new research protocols to the phones of the subjects. This way they can (for now) at least get the research questions from the same source without having to hard-code them in the app.