Formated all files with Prettier
In order to make future commits cleaner, I have formatted all the code that I had written with Prettier (https://github.com/prettier/prettier). The settings that I have changed from the default were the tabs to be 4 spaces, babylon as parser and to prefer double quotes to single quotes. All the rest of the settings were the same as the default.