Created Components out of Question Screens
There was a lot of repeating code when I was writing the question screens as separate components. So now there is a component for question, and components for each type of the question. This way, a lot of the code, that was shared between different types of questions is no longer duplicated.

This was the first step, to lower the number of places, where data needs to be encrypted or decrypted.