Notifications
I added the ability to show and schedule notifications for beeps. So far, it has only been tested on android. Both the date-pickers (needed to define the time range in which notifications are created) and notifications have different implementations on android and ios.

So far, the notifications are scheduled in the following way. The scheduler gets the end date and the number of notifications in this time. Then in randomly chooses the amount of times from now to end date and schedule notification at that time. It does not take any care for sleeping time. After the notification is taped, the app is opened, if on background, it comes forward and if it already opened, nothing changes.

As far as the date-pickers go, I used the react-native library, TimePickerAndroid and DatePickerAndroid for android and DatePickerIOS for iOS. The android version works. The iOS version renders the screen on android, but it tells me that this element is not available on android.

Next are notifications. The iOS uses the react-native library, but it needs to link the library to the project first. The process is described in https://facebook.github.io/react-native/docs/pushnotificationios.html#content with link to https://facebook.github.io/react-native/docs/linking-libraries-ios#manual-linking .

Android does not have a react-native option of implementation, so I used react-native-push-notifications. But in order to make the notifications also work on Androids 8.0 and higher, there needs to be a line changed in the react-native-push-notification package. The problem and the change is described in the pull request for the package: zo0r/react-native-push-notification#939 . It is just one line change.

Added dependencies: react-native-push-notifications: https://github.com/zo0r/react-native-push-notification