Location
Added the ability to use location in the app. The location can be used on two places. Once, the location can be saved as an answer to the question in the beep. The other one is, that app can keep track of the location, and when in the certain distance from the location, it sends a notification for a beep with a certain probability.

The code was not tested on iOS yet. There may be problems with permissions there. Check the following website in case of problems: https://facebook.github.io/react-native/docs/geolocation . Also, the notification when at location was also not tested yet.

The UI for the location as the answer needs to be rethought. The same is true for which information do we need for the location (out of the ones offered) or think about how to write the representation as recursion - but then the export in csv will be a bit more complicated.

Not sure, if the separate screen for permissions is needed.