Camera
Added the ability to take a picture. The feature was tested in a way, that it builds and it does not throw errors, but it was not tested in a way, that it works.

For iOS, it still needs the right permissions. Follow this link: https://github.com/react-native-community/react-native-image-picker/blob/master/docs/Install.md

Library used: react-native-image-picker https://github.com/react-native-community/react-native-image-picker