UI - Buttons, Settings, Beeps, Splash Screen
This commit is a couple of changes to the UI. The settings are now under the settings button in the footer on the home screen. I also added the splash screen for the time when the app is figuring out, if a person encrypted their data or not - this allows one less click for people that don't encrypt data. I also changed the beep representation, so the list of beeps opens a new screen with more information about the beep. I also changed how the buttons and tags look like. I took some of the colors from the images of the previous app, and for the others I tired to stick to the same pallet.