Email
Added the ability to send the generated csv files of beeps by email. I put all the functions connected with files to one screen and changed the UI (removing the password after submitting it, for example). Now the screen also saves the encrypted file upon opening, and deletes the files upon clicking the back button. There is still the ability to create file with encrypted content and send this instead.

There are two version of sending email. One does not use additional libraries, but uses the content of a csv in the body of email (which can be a problem because of line-breaks). The other one sends a csv as attachment, but uses the react-native-mail library. 

I also changed how the csv is structured, mostly by adding the header.

Added libraries: react-native-mail