App Icon
I added the icons for the app. On android, it shows at least an an icon to tap to open a program.

I added the bash script in the scripts to help create the icons for both platforms and puts them in the right place (folder: scripts). I found out the place and dimensions based on the generator-rn-toolbox library ( https://github.com/bamlab/generator-rn-toolbox/tree/master/generators/assets ), but I did not use the library itself.

I did not test it on iOS itself. If this way does not work, then follow the instructions in this StackOverflow answer: https://stackoverflow.com/a/34333915