Data Flow
In order to make a more consisted code, I had put the main functions
from reading and updating the storage in the seperate file. I did this,
in order to make the non-encryption and encryption mode work more
consistently and to have an easier job updating it, if some better
standards emerge (or I find out about it).

This also means that now all data is encrypted (except for password,
which is hashed), when before only beeps were encrypted. Because of
this, I also had to changed the way the id of new questions is calculated
(before I checked it in the database, now it picks the first integer
bigger than 1, which is still avalable).

While doing this, I also eliminated some code, which was not used. The
most prominent example is in the list of beeps. I also checked that now
all the screens also work with encryption.