User Experience
I changed and simplified the flow of different screens and (hopefully) with that simplified the user experience. I added all the questions on one screen and added the ability to have a dedicated place for experience description. I also made sure the pictures that are going to be taken can be added (so far manually) in email. I combined the files dealing with research and added the screen for API results of researches that are available. I also added the ability to change personal information and removed (but not deleted, since I will use these functions later) the screen with notifications.