UI Changes and Encryption
I repaired the Tags questions, as it took the possible tags from all the researches with this question id, now it also checks, if the tag was used in this research.

I made some changes to the CSV. One thing now is, that every line ends with |||, because the decryption method eats away the new lines.

Now the data, when send by email, can also be encrypted. The backup send to the email of a research participant can be encrypted by a password chosen in the screen in personal data. If password is deleted from there, then the data is no longer encrypted when sent.

The data for researcher (CSV) can be encrypted by choosing a share option when choosing a research and then picking the password. This password would need to be manually shared with the researcher. It is also now only possible to sent the data to researcher, if picking share and it would send only the data from their research.

There is also a html page to help researcher decrypt the data.