Notifications
Now, when the research is added or changed, the notifications and deleted and created anew. The person can also add the notifications for days they pick and remove the days that they don't want in personal settings.

I also put all the functions for notifications in the notification.js file.