Creating Testing Screens
I created the screen at the beginning, warning people that this is still
a testing version of the app and it can have problems and bugs and can
cause the loss of information. I also added the ability to send the
developer (me) email about the app in the settings.