Change of Research Plans
I added a way for people to change their research plans, in order for
them to be able to collect beeps for more than one research. People can
see all of their plans, which they started in the past, and they can
pick any one of them.

I also allows people to change their research plan without internet, as
before, it was only possible to start a new plan and for this, the plan
had to come from the internet.