Home Screen
I put the new beep on the home screen, in order to eliminate some of the
clicks, that people need to make, when they enter the app. Now the app
simply assumes, that if person leaves the app or they go to settings,
that they finished the beep and it saves it.