Changes in the first screen
There were some UX changes in the first screen.

The first one was, when the user changed the research question in the settings, it did not refresh to the new questions, until the user exited the app. Now it refreshes to the new questions, even when the user arrives via the back button.

The second one was reparing the bug, where the footer would not show, if
there were too many questions. Now it always shows.

The third was to put the camera and the main description of experience
in the questions area, as they are more similar to the questions about
experience than to the saving action.