Changed the UX Structure
I changed the UX structure. The analsis and data are now seperate from
data backup/sending and settings. In the analysis, I have changed the
structure of the beeps, to make them easier to read. In the setting, the
hours are now able to be picked with the UX (not handwritten) and now
there is only one screen connected with research questions. There are
also some new icons.