Notifications - Stop + Spacing
Now there is a button in settings that deletes all the already ceated
notifications. I also added the indication in what times will the
notifications be created.

Beside this, I also added the spacing between notifications. There is at
least 15 minutes now between two notifications (based on creation time).
If the time interval is too short, then the number of notifications will
be lower than the ones specified in the research plan.